APR.Promise, APR.Template and some minor fixes
This version fixes some minor problems in APRElement, APREvent and APRState.

It contains a new version of APRTemplate ready to be debugged, but since a lot
of my methods were implemented in a synchronous way (in all my code), I added
APR.Promise to try to fix that problem.

APR.Promise is ready to be used: it works, but it could work better.