BUG: Fixes bundles for testing and removes cross-browser testing functionality
This commit replaces Testling with Karma and JsDom. Cross-browser testing alternatives will be added soon because Testling didn't work at all.

DEP: Tape was replaced by Tap, but was deprecated in favor of Tape again

Tap seems to have better support and being more actively supported than Tape but loses the point of testing on the client side (and because Browserify didn't work when trying to bundle the client script [giving the error: "Cannot find the module: async_hooks"]).