ENH: Add support for element attributes in Define.urls
Allow passing an object in Define.urls to add custom attributes
in the loaded element.

Set the element's url to the object's "src" or "href" properties, if possible.
Check the "src" or "href" attribute to set the type of element.
Set the given attributes to the element before appending it. Ignore "href"
and "src" attributes.

DOC: Update docs accordingly

Add examples.

TST: Add unit test