BUG: Fix. Global is not being defined if a similar script is loading it
This bug appears to be consistent on webkit browsers. Neither JSDOM nor
mozilla seem to be affected.
Basically, if an <script> contains a global and you load it via html,
the first time you visit your website, it won't define that global
in just.Define.

The steps to reproduce this bug are the following:
- Create a file called index.html and put the following on it (remember to
change the path of the justjs file):
<html>
<head>
    <title>Bug</title>
</head>
<body>
    <script
        src='/path/to/just.js'
        data-just-Define='{"index": "index.js"}'
        async/></script>
    <script
        src='bug.js'
        data-just-Define='{"bug": "[src]"}'
        async></script>
</body>
</html>
- Create a file called "bug.js" and put "window.someGlobal = 1;" on it.
- Create another file called "index.js" and put the following on it:
  if ('someGlobal' in window) { just.Define('someGlobal', window.someGlobal); }
  else { just.Define.globals['someGlobal'] = function () { return
window.someGlobal; }; }
just.Define(['someGlobal'], function (someGlobal) { console.log(someGlobal); });
- Open index.html using a webkit browser.
- Open the console and reload the page (clear the cache if necessary).

After these steps, the console should log 1.

In order to fix this:

Add the same listener (as the created element) to similar element
and remove it when done. Client should define a global the right way, though.