BUG: Fix. Custom data is being ignored on Router#trigger()
e.detail is begin defaulted to custom object.
To fix it, this commit uses the "data" property on the emitted e.detail
to add custom data.
Also, it adds a 3rd argument on Router#trigger() to pass custom data
as the 2nd argument and eventInit on the 3rd one.

API: Set "data" as the 2nd argument for Router#trigger()

Pass custom data to Router#route() listener on e.detail.data property.

TST: Add unit tests for #trigger()