BUG: Fix. Router#change() is triggering eventInit as the custom data
Custom data is not taken into account in Router#change().

This commit adds an argument to Router#change() to accept custom data.
Also, it updates Router#push() and Router#replace() to allow passing
custom data.

API: Add a 4th argument to Router#change()

Pass data on the 3rd argument and eventInit on the 4th.
Update Router#push()
Update Router#replace()

API: Add a 3rd argument to Router#push()

Pass data on the 2rd argument and eventInit on the 3rd.

API: Add a 3rd argument to Router#replace()

Pass data on the 2nd argument and eventInit on the 3rd.

TST: Add tests for Router#change(), Router#push() and Router#replace()