BUG: Fix. Throttle "popstate" event when history.hash changes
When the History API is not supported, it fallbacks to changing
location.hash. The problem is that "location.hash" seems to trigger
a "popstate" event and history{pushState, replaceState} doesn't. That
causes handlers to be called before time.

This commit tries to fix that by throttling "popstate" events when
history.pushState is not available and the given paths check for a
hashbang (#!) in the url.