BUG: Fix. Remove elements on View.updateLoops() if some property is empty
Let's say we have the following element:
<div>
  <span
    data-var-for='value in values as data-non-empty'
    data-var='${value}'
    class='template'
    hidden></span>
</div>
And we call View.updateLoops() with some data, like {values: [1, 2]}.

The idea is that the span element will be cloned 2 times, with 1 and 2
as values (in this case, with each value as text). Then, if we set the
data to {values: null}, all the span elements must be removed,
except for the .template.

Before this commit, the view won't be updated if some property is empty,
causing the elements to don't be removed and preserve the previous state.

Hopefully, with this commit, elements get updated and removed as they should.