BUG: Fix. Access to any property on View~access()
Don't use regexps to check if it's a valid property. Just check for the
existence in the given object (using `in`). Eg. "a.#b" will return
undefined because "#b" is not a valid property/var. This should fix that.

This will break some tests but to fix them, check if the given data
is even an object. Otherwise return undefined, as expected by tests.
Also, remove View~isReservedKeyword() and View~isVar() as they're no longer
required.