BUG: Fix. Handle disconnected elements in View.updateLoops()
When an element is not connected to the DOM, but contains a cached
view, View.updateLoops() will throw if given as an element,
because its parentNode is undefined.

This fixes that by using the template's parentNode instead, only for
elements that have a cached view.

TST: Add unit test

Update test/jest/unit/lib/View.test.js.